Fix for edit class bug
